+++ py_src/snex/__init__.py
@@ -1,11 +1,13 @@
 from .etf import set_custom_encoder
-from .interface import call, cast, send
+from .interface import BeamModuleProxy, Elixir, call, cast, send
 from .logger import LoggingHandler
 from .models import Atom, DistinctAtom, EncodingOpt, EncodingOpts, Term
 
 __all__ = [
     "Atom",
+    "BeamModuleProxy",
     "DistinctAtom",
+    "Elixir",
     "EncodingOpt",
     "EncodingOpts",
     "LoggingHandler",